% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_football_uk.R
\name{clean_football_uk}
\alias{clean_football_uk}
\title{clean_football_uk}
\usage{
clean_football_uk(data = NULL)
}
\arguments{
\item{data}{dataset from the function "get_football_uk"}
}
\value{
tibble
}
\description{
cleaner for football_uk output.
}
