% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_int_results}
\alias{download_int_results}
\title{Download national team results}
\usage{
download_int_results()
}
\value{
tibble.
}
\description{
Results are downloaded from: https://github.com/martj42/international_results
}
\examples{
download_int_results()
}
