% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_call_API}
\alias{fdo_call_API}
\title{Call football-data.org API}
\usage{
fdo_call_API(request, token = "", response = "full")
}
\arguments{
\item{request}{(string) url fragment that completes the address of the
intended API call (required)}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
The output is converted from its JSON format to a data frame
}
\description{
Calls the football data.org API
}

