% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_getFixture}
\alias{fdo_getFixture}
\title{Show one fixture (including past meetings).}
\usage{
fdo_getFixture(id, past = "", token = "", response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "149461") of the fixture. Requried. No
default.}

\item{past}{(sting)Define the the number of former games to be analyzed in
the head2head node. Optional. Defaults to 10.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A list of two data frames (one list for the specified fixtures,
  another for the past meetings) converted from the JSON format is returned.
}
\description{
Returns information for a fixture (including past meetings) from the
football-data.org API
}
\examples{
\dontrun{
# fixture id for Athletic Bilbao v Sevilla  is 153986
# get fixture and past meetings information
fdo_getFixture("153986")
# just fixture information
fdo_getFixture("153986", past = 0)
}
}

