% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\docType{data}
\name{fdo_leagueCodes}
\alias{fdo_leagueCodes}
\title{fdo_leagueCodes}
\format{A data frame with 3 variables: \describe{
  \item{\code{leagueCode}}{The League Code as a string e.g. "PL" represents
  Premier League} \item{\code{country}}{The country in which that league
  takes place e.g. "England"} \item{\code{league}}{The name of the league
  e.g. "Premier League", "1.Bundesliga"} }

  For further details, see
  \url{http://api.football-data.org/docs/v1/index.html#league_codes}}
\usage{
fdo_leagueCodes
}
\description{
Table of leagueCodes to be used with \code{\link{fdo_listFixtures}}
}
\keyword{datasets}

