% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_leagueTable}
\alias{fdo_leagueTable}
\title{Competition league table from the football-data.org API}
\usage{
fdo_leagueTable(id, matchDay = "", token = "", response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "436") of the league in which the teams
compete. Requried. No default.}

\item{matchDay}{(string) The matchday to which the league table applies
(e.g. "3"). Optional. Default is the current match day.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns a competition league table from the football-data.org API
}
\examples{
\dontrun{
# get current Spanish Primera Division (La Liga) standings (without metadata)
fdo_leagueTable("436",response = "minified")
# Primera Division league table afer match day 2
fdo_leagueTable("436",response = "minified",matchDay = "2")
}
}

