% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_listCompFixtures}
\alias{fdo_listCompFixtures}
\title{List all fixtures for a certain competition.}
\usage{
fdo_listCompFixtures(id, timeFrame = "", matchDay = "", token = "",
  response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "436") of the competition. Requried. No
default.}

\item{timeFrame}{(string) The value of the timeFrame argument must start
with either p(ast) or n(ext), representing a timeframe either in the past
or future. It is followed by a number in the range 1-99. For example, p6
would return all fixtures in the last 6 days, whereas n23 would result in
returning all fixtures in the next 23 days. Optional. Default is the empty
string, which returns all fixtures, irrespective of date.}

\item{matchDay}{(string) The matchday to which the fixtures relate (e.g.
"3"). Optional. Default is "", which returns all matchDays. Note that a
non-empty string matchDay value will override the timeFrame value.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns all fixtures for a specific competition from the football-data.org
API
}
\examples{
\dontrun{
# league id for Ligue 1 2015/16 is 434
# all fixtures in the Ligue 1 2015/16 season
fdo_listCompFixtures("434")
# Ligue 1 fixtures in the next 10 days
fdo_listCompFixtures("434",timeFrame="n10")
# Ligue 1 fixtures from the previous 10 days
fdo_listCompFixtures("434",timeFrame="p10")
# Ligue 1 fixtures for match day 10
fdo_listCompFixtures("434",matchDay="10")
}
}

