% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_listComps}
\alias{fdo_listComps}
\title{List of competitions from the football-data.org API}
\usage{
fdo_listComps(season = "", token = "", response = "full")
}
\arguments{
\item{season}{(sting) The season (e.g. "2015") to which the competitions
relate. Note that the convention is to take the starting year to indicate
the season. For example, "2015/2016" is represented by "2015". Defaults to
the current season.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns list of competitions from the football-data.org API
}
\examples{
\dontrun{
# defaults to current season
fdo_listComps()
# get list of 2015 competitions without metadata
fdo_listComps("2015",response = "minified")
}
}

