% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_listCompTeams}
\alias{fdo_listCompTeams}
\title{List of teams in competitions from the football-data.org API}
\usage{
fdo_listCompTeams(id, token = "", response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "436") of the league in which the teams
compete. Requried. No default.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default is
"full" and "minified" will remove some meta inforation and thus return a
smaller data frame. Finally, "compressed" is only supported by the fixture
resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns list of teams in a competition from the football-data.org API
}
\examples{
\dontrun{
# get list of teams (without metadata) competing in the Spanish Primera Division (La Liga)
fdo_listCompTeams("436",response = "minified")
}
}

