% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_listFixtures}
\alias{fdo_listFixtures}
\title{List all fixtures for a certain competition.}
\usage{
fdo_listFixtures(league = "", timeFrame = "", token = "",
  response = "full")
}
\arguments{
\item{league}{(string <vector>) A vector of the league codes. Not to be
confused with league ids, leagues codes can be found the
\code{\link{fdo_leagueCodes}} data frame. Optional. Default is "", meaning
all league codes are returned.}

\item{timeFrame}{(string) The value of the timeFrame argument must start
with either p(ast) or n(ext), representing a timeframe either in the past
or future. It is followed by a number in the range 1-99. For example, p6
would return all fixtures in the last 6 days, whereas n23 would result in
returning all fixtures in the next 23 days. Optional. Default is "n7".}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns all fixtures for a specific competition from the football-data.org
API
}
\examples{
\dontrun{
# league code for La Liga is "PD"
# all La Liga fixtures in the coming 7 days
fdo_listFixtures("PD")
# league code for Ligue 1 is "FL1"
# all La Liga and Ligue 1 fixtures in the next 14 days.
fdo_listFixtures(league=c("PD","FL1"),timeFrame="n14")
}
}

