% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_listTeamFixtures}
\alias{fdo_listTeamFixtures}
\title{List all fixtures for a certain team.}
\usage{
fdo_listTeamFixtures(id, season = "", timeFrame = "", venue = "",
  token = "", response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "77") of the team Requried. No default.}

\item{season}{(sting) The season (e.g. "2015") to which the competitions
relate. Note that the convention is to take the starting year to indicate
the season. For example, "2015/2016" is represented by "2015". Optional.
Defaults to the current season.}

\item{timeFrame}{(string) The value of the timeFrame argument must start
with either p(ast) or n(ext), representing a timeframe either in the past
or future. It is followed by a number in the range 1-99. For example, p6
would return all fixtures in the last 6 days, whereas n23 would result in
returning all fixtures in the next 23 days. Optional. Default is the empty
string, which returns all fixtures for the current season.}

\item{venue}{(sting) Determines whether home, away or all fixtures are
returned. Possible values are "home", "away" and "". Optional. Default is
"", which means all fixtures are returned.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns all fixtures for a specific team from the football-data.org API
}
\examples{
\dontrun{
# team id for Athletic Bilbao is 77
# all Athletic Bilbao fixtures in the current (2016/17) season
fdo_listTeamFixtures("77")
# all Athletic Bilbao fixtures in 2015/16 season
fdo_listTeamFixtures("77",season = "2015")
# all Athletic Bilbao fixtures in the next 10 days
fdo_listTeamFixtures("77",timeFrame="n10")
# all Athletic Bilbao home fixtures in the next 10 days
fdo_listTeamFixtures("77",timeFrame = "n10", venue = "home")
}
}

