% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdo.R
\name{fdo_teamPlayers}
\alias{fdo_teamPlayers}
\title{Show all players for one team}
\usage{
fdo_teamPlayers(id, token = "", response = "full")
}
\arguments{
\item{id}{(sting) The id (e.g. "77") of the team. Requried. No default.}

\item{token}{(string) football data.org API token. Optional but it's
recommended that you acquire an API token. It allows the owner to document
API activity. See \url{http://api.football-data.org/register} for more
details.}

\item{response}{(string). Determine the structure of the response. Default
is "full" and "minified" will remove some meta inforation and thus return
a smaller data frame. Finally, "compressed" is only supported by the
fixture resource.}
}
\value{
A data frame converted from the JSON format is returned.
}
\description{
Returns player data for a specific team from the football-data.org API
}
\examples{
\dontrun{
# the team id for Athletic Bilbao is 77
# return data for all Athletic Bilbao players
fdo_teamPlayers("77")
}
}

