% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{football}
\alias{football}
\title{football}
\usage{
football(country)
}
\arguments{
\item{country}{the name of the country you want to import}
}
\description{
funzione per importare tutti i dati di un paese
}
