% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_anni.R
\name{football_anni}
\alias{football_anni}
\title{football_anni}
\usage{
football_anni(country, div, start_year = 2003, end_year = 2019)
}
\arguments{
\item{country}{country}

\item{div}{division}

\item{start_year}{anno di inizio}

\item{end_year}{anno di fine}
}
\description{
funzione per importare tutti gli anni di ua divisione
}
