% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_country.R
\name{football_country}
\alias{football_country}
\title{football_country}
\usage{
football_country(country)
}
\arguments{
\item{country}{country}
}
\description{
funzione per importare tutti i dati di un paese
}
