% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\docType{data}
\name{football_data}
\alias{football_data}
\title{football}
\format{An object of class \code{list} of length 10.}
\usage{
football_data
}
\arguments{
\item{save}{TRUE / FALSE}

\item{name}{newname}
}
\description{
football
}
\keyword{datasets}
