% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_.R
\name{football_single}
\alias{football_single}
\title{football_}
\usage{
football_single(country = NULL, div = "div1", year = "2018",
  raw = FALSE, quiet = TRUE)
}
\arguments{
\item{country}{country}

\item{div}{division}

\item{year}{year}

\item{raw}{TRUE / FALSE}

\item{quiet}{TRUE /FALSE}
}
\description{
"rh"
}
