% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{football_uk}
\alias{football_uk}
\title{football_uk}
\usage{
football_uk(
  country = "all",
  division = "all",
  years = c(start = 2010, end = 2020),
  quiet = FALSE
)
}
\arguments{
\item{country}{character vector, one or more of the available countries.
By default is set to 'all' and will consider all the countries available.}

\item{division}{character vector, one or more of the available divisions.
By default is set to 'all' and will consider all the divisions available for a specified country.}

\item{years}{integer vector, the first entry is the first year for importation while the second entry the last year.}

\item{quiet}{logical, when TRUE function evaluates without displaying customary messages.}
}
\value{
tibble
}
\description{
A wrapper for the data available on \href{https://www.football-data.co.uk}{football-uk}.
}
