% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{footballdatauk-deprecated}
\alias{footballdatauk-deprecated}
\alias{football_uk}
\alias{get_football_uk}
\alias{clean_football_uk}
\title{Deprecated function names}
\usage{
football_uk(
  country = "all",
  division = "all",
  years = c(2020, 2024),
  quiet = FALSE,
  verbose = NULL,
  ...
)

get_football_uk(
  country = NULL,
  division = "div1",
  year = 2020,
  raw = FALSE,
  quiet = FALSE,
  verbose = NULL,
  ...
)

clean_football_uk(data = NULL, year = NULL, ...)
}
\arguments{
\item{country, division, years, year, raw, quiet}{Arguments from the old API.}

\item{verbose}{Deprecated. Use \code{quiet}; \code{verbose = TRUE} becomes
\code{quiet = FALSE}.}

\item{...}{Arguments passed to the replacement function.}

\item{data}{A data frame or tibble created from a football-data.co.uk CSV.}
}
\value{
A tibble.
}
\description{
These functions are retained as soft-deprecated wrappers around the \verb{ftuk_*}
API.
}
