% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{footballdatauk-package}
\alias{footballdatauk}
\alias{footballdatauk-package}
\title{footballdatauk: Read and Clean Football Results}
\description{
Reads and downloads football match results and betting odds from football-data.co.uk and converts them to a consistent tibble schema. The package also includes a small local Italian Serie A example dataset for documentation, tests, and offline examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beniamino98/footballdatauk}
  \item Report bugs at \url{https://github.com/beniamino98/footballdatauk/issues}
}

}
\author{
\strong{Maintainer}: Beniamino Sartini \email{sartinibeniamino@gmail.com}

}
\keyword{internal}
