% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ftuk_countries}
\alias{ftuk_countries}
\title{Supported football-data.co.uk countries}
\usage{
ftuk_countries()
}
\value{
A character vector of supported country identifiers.
}
\description{
Supported football-data.co.uk countries
}
\examples{
ftuk_countries()

}
