% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ftuk_divisions}
\alias{ftuk_divisions}
\title{Supported football-data.co.uk divisions}
\usage{
ftuk_divisions(country)
}
\arguments{
\item{country}{Character scalar. A supported country.}
}
\value{
A tibble with \code{country}, \code{division}, \code{code}, \code{label}, \code{first_year},
and \code{last_year} columns.
}
\description{
Supported football-data.co.uk divisions
}
\examples{
ftuk_divisions("italy")

}
