% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{ftuk_seasons}
\alias{ftuk_seasons}
\title{Supported football-data.co.uk seasons}
\usage{
ftuk_seasons()
}
\value{
A tibble with season-ending \code{year}, \code{season}, and football-data.co.uk
\code{code}. For example, \code{year = 2024} means season 2023-2024.
}
\description{
Supported football-data.co.uk seasons
}
\examples{
tail(ftuk_seasons())

}
