% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-download.R
\name{ftuk_url}
\alias{ftuk_url}
\title{Build a football-data.co.uk CSV URL}
\usage{
ftuk_url(country, division, year)
}
\arguments{
\item{country}{Character scalar. A supported country.}

\item{division}{Character scalar. A division supported for \code{country}.}

\item{year}{Numeric scalar. The season-ending year. For example,
\code{year = 2024} means season 2023-2024.}
}
\value{
A character scalar URL.
}
\description{
Build a football-data.co.uk CSV URL
}
\examples{
ftuk_url("italy", "div1", 2024)

}
