% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{get_divisions}
\alias{get_divisions}
\title{get_divisions}
\usage{
get_divisions(country = NULL, division = "all", year = c(start =
  2010, end = 2020), quiet = FALSE)
}
\arguments{
\item{country}{character. One of the available countries. See  \code{info_country}.}

\item{division}{character vector. One or more of the available divisions. See  \code{info_divisions}. 
By default is set to "all" and will consider all the divisions availables for a specific country.}

\item{year}{integer vector. The first item is the first year of importation and the second item is the last year.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data availables on \href{https://www.football-data.co.uk}{football-uk}.
}
