% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_football_uk.R
\name{get_football_uk}
\alias{get_football_uk}
\title{get_football_uk}
\usage{
get_football_uk(
  country = NULL,
  division = "div1",
  year = 2020,
  raw = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{country}{character. one of the available country.}

\item{division}{character, one of the available divisions.}

\item{year}{character, the season year, for example for season 2022/2023 the year will be 2023.}

\item{raw}{logical, FALSE as default. If TRUE the function will return a tibble as it is stored on \href{https://www.football-data.co.uk}{football-uk},
otherwise it will clean and structure the result.}

\item{quiet}{logical, when TRUE function evaluates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to download and clean the data available on \href{https://www.football-data.co.uk}{football-uk}.
}
