% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_webdata.R
\name{get_webdata}
\alias{get_webdata}
\title{get_webdata}
\usage{
get_webdata(country = NULL, division = "div1", year = 2020,
  raw = FALSE, quiet = FALSE)
}
\arguments{
\item{country}{character. One of the available countries. See  \code{info_country}.}

\item{division}{character. One of the available divisions See  \code{info_divisions}.}

\item{year}{character. The year for which you want to import the data.}

\item{raw}{logical. If TRUE will be applied a function to reduce and rename the variables, 
otherwise it will return the original data.frame.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data availables on \href{https://www.football-data.co.uk}{football-uk}.
}
