% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{get_years}
\alias{get_years}
\title{get_years}
\usage{
get_years(country = NULL, division = "div1", year = c(start = 2010,
  end = 2020), quiet = FALSE)
}
\arguments{
\item{country}{character. One of the available countries. See  \code{info_country}.}

\item{division}{character. One of the available divisions See  \code{info_divisions}.}

\item{year}{integer vector. The first item is the first year of importation and the second item is the last year.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data availables on \href{https://www.football-data.co.uk}{football-uk}.
}
