% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_divisions}
\alias{info_divisions}
\title{info_divisions}
\usage{
info_divisions(country = "all")
}
\arguments{
\item{country}{: character. One of the available countries. See  \code{info_country}}
}
\description{
a function to obtain all the available division for all countries or for a specific country.
}
