% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_country}
\alias{is_country}
\title{is_country}
\usage{
is_country(country = NULL)
}
\arguments{
\item{country}{: character. One of the available countries. See  \code{info_country}}
}
\description{
a function to check if is a valid country or not.
}
