% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_division}
\alias{is_division}
\title{is_division}
\usage{
is_division(country = NULL, division = NULL)
}
\arguments{
\item{country}{: character. One of the available countries. See  \code{info_country}}

\item{division}{: character. One of the available divisions See  \code{info_divisions}}
}
\description{
a function to check if is a valid division or not.
}
