% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_if}
\alias{na_if}
\title{na_if}
\usage{
na_if(x = NULL, .if = c("", " ", "N/A", "NA"))
}
\arguments{
\item{x}{an object, can be a vector, a list or a dataframe.}

\item{.if}{others type of NA's. By default is  c("", " ", "N/A", "NA")}
}
\description{
a function to substitute others forms of NA's. It is a consistent function in the sense 
            that if x is a list the output will be a list;
            if is a data.frame, will be a data.frame;
            and if it is a vector it will be a vector.
}
