% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_read_data.R
\name{PL_read_data}
\alias{PL_read_data}
\title{Function to read in Premier League data}
\usage{
PL_read_data(from, to)
}
\arguments{
\item{from}{Start date}

\item{to}{End date}
}
\description{
This function retrieves data for Premier League matches inbetween the inputed dates.
}
