% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.R
\name{Team}
\alias{Team}
\title{Team}
\usage{
Team(.data, team = NULL, year = NULL, vs = NULL, away = FALSE,
  home = FALSE)
}
\arguments{
\item{.data}{the data}

\item{team}{character the team}

\item{year}{vector, the years}

\item{vs}{vector, the other teams}

\item{away}{logical, if TRUE all match away}

\item{home}{logical, if TRUE all match in home}
}
\description{
filter data
}
