% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_cols}
\alias{team_cols}
\title{Premier League team names and team colours}
\format{
A data frame with 41 rows and 5 variables:
\describe{
  \item{teams}{Character for name of years}
  \item{col1}{Teams first colour}
  \item{col2}{Teams second colour}
  \item{col1_rgb}{RGB version of first colour}
  \item{col2_rgb}{RGB version of second colour}
  }
}
\usage{
team_cols
}
\description{
A dataset containing the teams active in the premier league
for the seasons from 2000 until 2018, and their team colours
}
\keyword{datasets}
