% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{year_dates}
\alias{year_dates}
\title{Premier League seasons start and end dates}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{years}{Character for name of years}
  \item{start_date}{Dates of the first match in each season}
  \item{end_date}{Dates of the last match in each season}
  }
}
\usage{
year_dates
}
\description{
A dataset containing the start and end dates of all premier league seasons from 2000 until 2018.
}
\keyword{datasets}
