% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_football_uk.R
\name{clean_football_uk}
\alias{clean_football_uk}
\title{clean_football_uk}
\usage{
clean_football_uk(x = NULL)
}
\arguments{
\item{x}{dataset from the function "get_football_uk"}
}
\value{
tibble
}
\description{
clean_football_uk
}
