% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{football}
\alias{football}
\title{football}
\usage{
football(country = "italy", div = "div1", from = 2003, to = 2019,
  raw = FALSE, quiet = FALSE)
}
\arguments{
\item{country}{country}

\item{div}{division}

\item{from}{anno di inizio}

\item{to}{anno di fine}

\item{raw}{jjs}

\item{quiet}{TRUE /FALSE}
}
\description{
funzione per importare tutti gli anni di ua divisione
}
