% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_single.R
\name{football_single}
\alias{football_single}
\title{football_single}
\usage{
football_single(country, div = "div1", anno = "2018", verbose = TRUE)
}
\arguments{
\item{country}{country}

\item{div}{division}

\item{anno}{year}

\item{verbose}{TRUE /FALSE}
}
\description{
football_single
}
