% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{football_uk}
\alias{football_uk}
\title{football_uk}
\usage{
football_uk(country = "all", division = "all", year = c(start = 2010,
  end = 2020), quiet = FALSE)
}
\arguments{
\item{country}{character vector. One or more of the availables countries. See  \code{info_country}.
By default is set to "all" and will consider all the countries availables.}

\item{division}{character vector. One or more of the availables divisions. See  \code{info_divisions}. 
By default is set to "all" and will consider all the divisions availables for a specific country.}

\item{year}{integer vector. The first item is the first year of importation and the second item is the last year.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data availables on \href{https://www.football-data.co.uk}{football-uk}.
}
