% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{get_divisions}
\alias{get_divisions}
\title{get_divisions}
\usage{
get_divisions(
  country = NULL,
  division = "all",
  years = c(start = 2010, end = 2020),
  verbose = TRUE
)
}
\arguments{
\item{country}{character. One of the available countries.}

\item{division}{character vector. One or more of the available divisions.
By default is set to 'all' and will consider all the divisions available for a specific country.}

\item{years}{integer vector. The first item is the first year of importation and the second item is the last year.}

\item{verbose}{logical. When TRUE function evaluates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data available on \href{https://www.football-data.co.uk}{football-uk}.
}
