% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_football_uk.R
\name{get_football_uk}
\alias{get_football_uk}
\title{get_football_uk}
\usage{
get_football_uk(
  country = NULL,
  division = "div1",
  year = 2020,
  raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{country}{character. One of the available countries.}

\item{division}{character. One of the available divisions.}

\item{year}{character. The year for which you want to import the data.}

\item{raw}{logical. If TRUE will be applied a function to reduce and rename the variables,
otherwise it will return the original data.frame.}

\item{quiet}{logical. When TRUE function evalueates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data availables on \href{https://www.football-data.co.uk}{football-uk}.
}
