% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_uk.R
\name{get_years}
\alias{get_years}
\title{get_years}
\usage{
get_years(
  country = NULL,
  division = "div1",
  years = c(start = 2010, end = 2020),
  verbose = TRUE
)
}
\arguments{
\item{country}{character. One of the available countries.}

\item{division}{character. One of the available divisions.}

\item{years}{integer vector. The first item is the first year of importation and the second item is the last year.}

\item{verbose}{logical. When TRUE function evaluates without displaying customary messages.}
}
\value{
tibble
}
\description{
base function to get in a clean way the data available on \href{https://www.football-data.co.uk}{football-uk}.
}
